+++ app/src/main.tsx
@@ -1,7 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { PostHogProvider } from "@posthog/react";
 import App from "./App";
 import "./index.css";
+import { posthogOptions, posthogProjectToken } from "./lib/posthog";
 import { initTheme } from "./lib/theme";
 
 // Apply stored/system theme BEFORE React mounts to avoid a color flash.
@@ -9,6 +11,8 @@ initTheme();
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <App />
+    <PostHogProvider apiKey={posthogProjectToken} options={posthogOptions}>
+      <App />
+    </PostHogProvider>
   </React.StrictMode>,
 );